feat: structured-packages-implementation-rec-05-artifacts-export-006 - Package-boundary evidence and delivery through merge
