+++ pkg/services/recordings/internal/services/recording_lifecycle/internal/service/active_flush_test.go
@@ -228,6 +228,7 @@ func newActiveFlushRoot(
 		nil,
 		writer,
 		tickers,
+		nil,
 	)
 }
