+++ pkg/services/recordings/internal/services/recording_lifecycle/internal/service/failure_accumulation_test.go
@@ -231,6 +231,7 @@ func newFailureTestRoot(
 		nil,
 		writer,
 		tickers,
+		nil,
 		clock,
 	)
 }
